removed all implementation for a re-start, now compiles
