+++ src/lib.rs
@@ -225,15 +225,19 @@ along with range_bounds_map. If not, see <https://www.gnu.org/licenses/>.
 #![feature(btree_cursors)]
 #![allow(clippy::tabs_in_doc_comments)]
 #![allow(clippy::needless_return)]
+#![allow(unused_imports)]
+#![allow(unused_variables)]
 pub(crate) mod bound_ord;
 pub(crate) mod custom_range_bounds_ord_wrapper;
+pub(crate) mod helpers;
+
 pub mod range_bounds_map;
-pub mod range_bounds_set;
+//pub mod range_bounds_set;
 pub mod try_from_bounds;
 
 pub use crate::range_bounds_map::{
 	OverlapError, OverlapOrTryFromBoundsError, RangeBoundsMap,
 	TryFromBoundsError,
 };
-pub use crate::range_bounds_set::RangeBoundsSet;
+//pub use crate::range_bounds_set::RangeBoundsSet;
 pub use crate::try_from_bounds::TryFromBounds;
